@@ -1,6 +1,6 @@
 {
   "name": "@studiometa/js-toolkit-workspace",
-  "version": "2.0.0-rc.2",
+  "version": "2.0.0",
   "private": true,
   "workspaces": [
     "packages/*"
@@ -22,19 +22,19 @@
     "build:pkg": "node scripts/build.js"
   },
   "devDependencies": {
-    "@studiometa/eslint-config": "^2.1.3",
+    "@studiometa/eslint-config": "^3.0.1",
     "@studiometa/prettier-config": "^2.1.1",
     "@studiometa/stylelint-config": "^2.0.0",
-    "@types/estree": "^0.0.50",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^16.11.13",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
-    "esbuild": "^0.14.12",
-    "eslint": "^7.32.0",
-    "eslint-plugin-jest": "^24.4.0",
+    "@types/estree": "^0.0.51",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.23",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
+    "esbuild": "^0.14.29",
+    "eslint": "^8.12.0",
+    "eslint-plugin-jest": "^26.1.3",
     "fast-glob": "^3.2.11",
-    "prettier": "^2.5.1",
-    "typescript": "^4.5.4"
+    "prettier": "^2.6.1",
+    "typescript": "^4.6.3"
   }
 }
