@@ -1,6 +1,6 @@
 {
   "name": "@studiometa/js-toolkit-demo",
-  "version": "2.0.0-rc.2",
+  "version": "2.0.0",
   "private": true,
   "type": "commonjs",
   "scripts": {
@@ -8,17 +8,17 @@
     "build": "meta build"
   },
   "dependencies": {
-    "@studiometa/eslint-config": "^2.1.3",
+    "@studiometa/eslint-config": "^3.0.1",
     "@studiometa/stylelint-config": "^2.0.0",
-    "@studiometa/tailwind-config": "^1.0.4",
-    "@studiometa/ui": "^0.1.0",
-    "@studiometa/webpack-config": "^3.0.0",
-    "eslint": "^7.32.0",
-    "postcss": "^8.3.6",
-    "prettier": "^2.3.2",
+    "@studiometa/tailwind-config": "^1.1.0",
+    "@studiometa/ui": "^0.2.3",
+    "@studiometa/webpack-config": "^3.1.1",
+    "eslint": "^8.12.0",
+    "postcss": "^8.4.12",
+    "prettier": "^2.6.1",
     "stylefire": "^7.0.3",
-    "stylelint": "^13.13.1",
-    "tailwindcss": "^2.2.9",
+    "stylelint": "^14.6.1",
+    "tailwindcss": "^3.0.23",
     "vue": "^2.6.14"
   }
 }
