@@ -1,14 +1,28 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-const { defineConfig } = require('vite');
-const path = require('path');
+import { defineConfig } from 'vite';
+import fs from 'fs';
+import path from 'path';
 
-module.exports = defineConfig({
+export default defineConfig({
   optimizeDeps: {
     include: ['@studiometa/js-toolkit'],
   },
+  plugins: [
+    {
+      name: 'add-common-js-package-plugin',
+      writeBundle(viteConfig) {
+        if (viteConfig.format === 'cjs' && viteConfig.esModule) {
+          fs.writeFileSync(
+            path.join(viteConfig.dir, 'package.json'),
+            JSON.stringify({ type: 'commonjs' })
+          );
+        }
+      },
+    },
+  ],
   resolve: {
     alias: {
-      './NavBarTitle.vue': path.resolve(__dirname, '.vitepress/theme/components/NavBarTitle.vue'),
+      './NavBarTitle.vue': path.resolve('.vitepress/theme/components/NavBarTitle.vue'),
     },
   },
 });
