@@ -1,4 +1,4 @@
 module.exports = {
   presets: [require('@studiometa/tailwind-config')],
-  purge: ['src/templates/**/*.twig', 'src/**/*.js'],
+  content: ['src/templates/**/*.twig', 'src/**/*.js'],
 };
