@@ -6,8 +6,8 @@
  * Test if an object has a method.
  *
  * @param  {Object}  obj  The object to test
- * @param  {String}  name The method's name
- * @return {Boolean}
+ * @param  {string}  name The method's name
+ * @returns {boolean}
  */
 export function hasMethod(obj, name) {
   return typeof obj[name] === 'function';
@@ -16,11 +16,11 @@ export function hasMethod(obj, name) {
 /**
  * Get a list of elements based on the name of a component.
  *
- * @param {String} nameOrSelector
+ * @param {string} nameOrSelector
  *   The name or selector to used for this component.
  * @param {HTMLElement|Document} element
  *   The root element on which to query the selector, defaults to `document`.
- * @return {Array<HTMLElement>}
+ * @returns {Array<HTMLElement>}
  *   A list of elements on which the component should be mounted.
  */
 export function getComponentElements(nameOrSelector, element = document) {
