@@ -1,16 +1,16 @@
 /**
  * @typedef {Object} Circle
- * @property {Number} x Circle's x position
- * @property {Number} y Circle's y position
- * @property {Number} radius Circle's radius
+ * @property {number} x Circle's x position
+ * @property {number} y Circle's y position
+ * @property {number} radius Circle's radius
  */
 
 /**
  * Convert clientRect to a formatted circle object
  *
  * @param {Partial<DOMRect>} domRect DOMRect of a square DOMElement
- * @param {Boolean} force Force usage of non-square DOMElements
- * @return {Circle} Circle object that can be used in collides functions
+ * @param {boolean} force Force usage of non-square DOMElements
+ * @returns {Circle} Circle object that can be used in collides functions
  */
 export default function boundingRectToCircle({ x, y, width, height }, force = false) {
   if (width !== height && !force)
