@@ -1,15 +1,14 @@
 /**
  * Format a CSS transform matrix with the given values.
  *
- * @param  {Object}  transform
- * @param  {Number=} [transform.scaleX=1]     The scale on the x axis.
- * @param  {Number=} [transform.scaleY=1]     The scale on the y axis.
- * @param  {Number=} [transform.skewX=0]      The skew on the x axis.
- * @param  {Number=} [transform.skewY=0]      The skew on the y axis.
- * @param  {Number=} [transform.translateX=0] The translate on the x axis.
- * @param  {Number=} [transform.translateY=0] The translate on the y axis.
- * @return {String}                           A formatted CSS matrix transform.
- *
+ * @param   {Object} transform
+ * @param   {number} [transform.scaleX=1]     The scale on the x axis.
+ * @param   {number} [transform.scaleY=1]     The scale on the y axis.
+ * @param   {number} [transform.skewX=0]      The skew on the x axis.
+ * @param   {number} [transform.skewY=0]      The skew on the y axis.
+ * @param   {number} [transform.translateX=0] The translate on the x axis.
+ * @param   {number} [transform.translateY=0] The translate on the y axis.
+ * @returns {string}                           A formatted CSS matrix transform.
  * @example
  * ```js
  * matrix({ scaleX: 0.5, scaleY: 0.5 });
