@@ -3,7 +3,7 @@ import clamp from './clamp.js';
 /**
  * Clamp a value in the 0–1 range.
  * @param {number} value
- * @return {number}
+ * @returns {number}
  */
 export default function clamp01(value) {
   return clamp(value, 0, 1);
