@@ -1,10 +1,10 @@
 /**
  * Interpolate the ratio between a given interval.
  *
- * @param  {Number} min   The interval minimum value.
- * @param  {Number} max   The inverval maximum value.
- * @param  {Number} ratio The ratio to get.
- * @return {Number}       The value between min and max corresponding to ratio.
+ * @param  {number} min   The interval minimum value.
+ * @param  {number} max   The inverval maximum value.
+ * @param  {number} ratio The ratio to get.
+ * @returns {number}       The value between min and max corresponding to ratio.
  */
 export default function lerp(min, max, ratio) {
   return (1 - ratio) * min + ratio * max;
