@@ -2,7 +2,7 @@
  * Wait for the next microtask.
  *
  * @template {() => any} T
- * @param    {T=} [fn]
+ * @param    {T} [fn]
  * @returns  {Promise<T extends Function ? ReturnType<T> : undefined>}
  */
 export default async function nextMicrotask(fn) {
