@@ -3,8 +3,9 @@ import withoutTrailingCharacters from './withoutTrailingCharacters.js';
 /**
  * Remove the given characters to the start of the given string recursively.
  *
- * @param {string} string     The string to modify.
- * @param {string} characters The characters to add to the start.
+ * @param   {string} string     The string to modify.
+ * @param   {string} characters The characters to add to the start.
+ * @returns {string}
  */
 export default function withoutTrailingCharactersRecursive(string, characters) {
   let str = withoutTrailingCharacters(string, characters);
